refactor: destroyRole
